move codeExamples into new directory
