+++ src/rules/classNameRule.ts
@@ -20,6 +20,7 @@ import * as ts from "typescript";
 
 import * as Lint from "../index";
 import { isUpperCase } from "../utils";
+import { codeExamples } from "./code-examples/className.examples";
 
 export class Rule extends Lint.Rules.AbstractRule {
     /* tslint:disable:object-literal-sort-keys */
@@ -32,22 +33,7 @@ export class Rule extends Lint.Rules.AbstractRule {
         optionExamples: [true],
         type: "style",
         typescriptOnly: false,
-        codeExamples: [
-            {
-                description: "Enforces PascalCased class and interface names.",
-                config: Lint.Utils.dedent`
-                    "rules": { "class-name": true }
-                `,
-                pass: Lint.Utils.dedent`
-                    class MyClass { }
-                    interface MyInterface { }
-                `,
-                fail: Lint.Utils.dedent`
-                    class myClass { }
-                    interface myInterface { }
-               `,
-            },
-        ],
+        codeExamples,
     };
     /* tslint:enable:object-literal-sort-keys */
