+++ src/rules/noAnyRule.ts
@@ -18,6 +18,7 @@
 import * as ts from "typescript";
 
 import * as Lint from "../index";
+import { codeExamples } from "./code-examples/noAny.examples";
 
 export class Rule extends Lint.Rules.AbstractRule {
     /* tslint:disable:object-literal-sort-keys */
@@ -31,20 +32,7 @@ export class Rule extends Lint.Rules.AbstractRule {
         optionExamples: [true],
         type: "typescript",
         typescriptOnly: true,
-        codeExamples: [
-            {
-                description: "Disallows usages of `any` as a type declaration.",
-                config: Lint.Utils.dedent`
-                    "rules": { "no-any": true }
-                `,
-                pass: Lint.Utils.dedent`
-                    let foo: object;
-                `,
-                fail: Lint.Utils.dedent`
-                    let foo: any;
-                `,
-            },
-        ],
+        codeExamples,
     };
     /* tslint:enable:object-literal-sort-keys */
